move the display of the board to a separate component
